requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
google-generativeai
websockets
passlib[bcrypt]
authlib
httpx
requests
pytest
pytest-asyncio
itsdangerous
starlette
